+++ pkg/controller/deployment/sync_test.go
@@ -32,7 +32,7 @@ import (
 )
 
 func intOrStrP(val int) *intstr.IntOrString {
-	intOrStr := intstr.FromInt(val)
+	intOrStr := intstr.FromInt(int32(val))
 	return &intOrStr
 }
