+++ pkg/proxy/ipvs/proxier_test.go
@@ -416,7 +416,7 @@ func TestExternalIPsNoEndpoint(t *testing.T) {
 				Name:       svcPortName.Port,
 				Port:       int32(svcPort),
 				Protocol:   api.ProtocolTCP,
-				TargetPort: intstr.FromInt(svcPort),
+				TargetPort: intstr.FromInt(int32(svcPort)),
 			}}
 		}),
 	)
@@ -470,7 +470,7 @@ func TestExternalIPs(t *testing.T) {
 				Name:       svcPortName.Port,
 				Port:       int32(svcPort),
 				Protocol:   api.ProtocolTCP,
-				TargetPort: intstr.FromInt(svcPort),
+				TargetPort: intstr.FromInt(int32(svcPort)),
 			}}
 		}),
 	)
@@ -709,7 +709,7 @@ func addTestPort(array []api.ServicePort, name string, protocol api.Protocol, po
 		Protocol:   protocol,
 		Port:       port,
 		NodePort:   nodeport,
-		TargetPort: intstr.FromInt(targetPort),
+		TargetPort: intstr.FromInt(int32(targetPort)),
 	}
 	return append(array, svcPort)
 }
