+++ test/e2e/common/autoscaling_utils.go
@@ -430,7 +430,7 @@ func runServiceAndWorkloadForResourceConsumer(c clientset.Interface, internalCli
 		Spec: v1.ServiceSpec{
 			Ports: []v1.ServicePort{{
 				Port:       port,
-				TargetPort: intstr.FromInt(targetPort),
+				TargetPort: intstr.FromInt(int32(targetPort)),
 			}},
 
 			Selector: map[string]string{
@@ -484,7 +484,7 @@ func runServiceAndWorkloadForResourceConsumer(c clientset.Interface, internalCli
 		Spec: v1.ServiceSpec{
 			Ports: []v1.ServicePort{{
 				Port:       port,
-				TargetPort: intstr.FromInt(targetPort),
+				TargetPort: intstr.FromInt(int32(targetPort)),
 			}},
 
 			Selector: map[string]string{
