+++ test/e2e/common/util.go
@@ -79,7 +79,7 @@ func svcByName(name string, port int) *v1.Service {
 			},
 			Ports: []v1.ServicePort{{
 				Port:       int32(port),
-				TargetPort: intstr.FromInt(port),
+				TargetPort: intstr.FromInt(int32(port)),
 			}},
 		},
 	}
